requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect==1.1.0
greenlet==1.1.3.post0
idna==3.4
lxml==4.9.1
numpy==1.23.4
pandas==1.5.0
parsel==1.6.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
urllib3==1.26.12
w3lib==2.0.1
